pyproject.toml
[project]
name = "transcribe-srt"
version = "0.1.0"
description = ""
authors = [
    {name = "miruohotspring",email = "miruohotspring@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=0.27.0,<0.28.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "transcribe_srt", from = "src"}]

[tool.poetry.scripts]
gen-sub = "transcribe_srt:main"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
